Ports configs/etc to PIC32MX795

+++ FT_DataTypes.h
@@ -1,7 +1,6 @@
 #ifndef _FT_DATATYPES_H_
 #define _FT_DATATYPES_H_
 
-
 #define FT_FALSE           (0)
 #define FT_TRUE            (1)
 
@@ -9,14 +8,12 @@ typedef unsigned char ft_uint8_t;
 typedef char ft_char8_t;
 typedef signed char ft_schar8_t;
 typedef unsigned char ft_uchar8_t;
-typedef int  ft_int16_t;
-typedef unsigned int ft_uint16_t;
+typedef short ft_int16_t;
+typedef unsigned short ft_uint16_t;
 typedef unsigned long ft_uint32_t;
 typedef long ft_int32_t;
 typedef void ft_void_t;
-
 typedef unsigned char ft_bool_t;
-
 
 #endif /*_FT_DATATYPES_H_*/
